Rename unified listing RSS source
