+++ BinanceUsdtTicker.csproj
@@ -25,8 +25,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Remove="KuCoinNewListingsRss.cs" />
-    <None Include="KuCoinNewListingsRss.cs" />
+    <Compile Remove="NewListingRss.cs" />
+    <None Include="NewListingRss.cs" />
   </ItemGroup>
 
 </Project>
